+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.1.27
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/btp-utils@1.0.2
+    -   @sap-ux/inquirer-common@0.6.29
+    -   @sap-ux/ui5-config@0.26.4
+    -   @sap-ux/deploy-config-generator-shared@0.0.37
+    -   @sap-ux/cf-deploy-config-inquirer@0.2.18
+    -   @sap-ux/cf-deploy-config-writer@0.1.22
+    -   @sap-ux/fiori-generator-shared@0.9.9
+    -   @sap-ux/project-access@1.29.16
+
 ## 0.1.26
 
 ### Patch Changes